fix: extract organization config to break server-client import chain
getAppOrganizationId and getAppOrganizationIdClient are pure env-var
getters with no server dependencies. Moving them to organization-config.ts
prevents utilities.ts (and its client consumers) from transitively
importing next/headers via server-client.ts.
